misc cleanups
