+++ dataplane/handlers/interface.go
@@ -43,7 +43,7 @@ type Interface struct {
 	c *ygnmi.Client
 	// closers functions should all be invoked when the interface handler stops running.
 	closers []func()
-	fwd     fwdpb.ServiceClient
+	e       *engine.Engine
 	stateMu sync.RWMutex
 	// state keeps track of the applied state of the device's interfaces so that we do not issue duplicate configuration commands to the device's interfaces.
 	state     map[string]*oc.Interface
@@ -65,9 +65,9 @@ type interfaceManager interface {
 }
 
 // NewInterface creates a new interface handler.
-func NewInterface(fwd fwdpb.ServiceClient) *reconciler.BuiltReconciler {
+func NewInterface(e *engine.Engine) *reconciler.BuiltReconciler {
 	i := &Interface{
-		fwd:       fwd,
+		e:         e,
 		idxToName: map[int]string{},
 		state:     map[string]*oc.Interface{},
 		ifaceMgr:  &kernel.Interfaces{},
@@ -187,10 +187,7 @@ func (ni *Interface) startCounterUpdates(ctx context.Context) {
 			}
 			ni.stateMu.RUnlock()
 			for _, intfName := range intfNames {
-				countersReply, err := ni.fwd.ObjectCounters(ctx, &fwdpb.ObjectCountersRequest{
-					ObjectId:  &fwdpb.ObjectId{Id: intfName},
-					ContextId: &fwdpb.ContextId{Id: engine.DefaultContextID},
-				})
+				countersReply, err := ni.e.GetCounters(ctx, intfName)
 				log.V(2).Infof("querying counters for interface %q, got %v", intfName, countersReply)
 				if err != nil {
 					log.Errorf("interface handler: could not retrieve counter for interface %q", intfName)
@@ -346,7 +343,7 @@ func (ni *Interface) handleLinkUpdate(ctx context.Context, lu *netlink.LinkUpdat
 
 	modelName := engine.TapNameToIntfName(lu.Attrs().Name)
 	iface := ni.getOrCreateInterface(modelName)
-	if err := engine.UpdatePortSrcMAC(ctx, ni.fwd, modelName, lu.Attrs().HardwareAddr); err != nil {
+	if err := ni.e.UpdatePortSrcMAC(ctx, modelName, lu.Attrs().HardwareAddr); err != nil {
 		log.Warningf("failed to update src mac: %v", err)
 	}
 	iface.GetOrCreateEthernet().MacAddress = ygot.String(lu.Attrs().HardwareAddr.String())
@@ -410,7 +407,7 @@ func (ni *Interface) handleAddrUpdate(ctx context.Context, au *netlink.AddrUpdat
 			gnmiclient.BatchUpdate(sb, ocpath.Root().Interface(modelName).Subinterface(0).Ipv6().Address(ip).PrefixLength().State(), uint8(pl))
 		}
 		// Forward all packets destined to this interface to the corresponding TAP interface.
-		if err := engine.AddLayer3PuntRule(ctx, ni.fwd, modelName, ipBytes); err != nil {
+		if err := ni.e.AddLayer3PuntRule(ctx, modelName, ipBytes); err != nil {
 			log.Warningf("failed to add layer3 punt rule: %v", err)
 		}
 	} else {
@@ -443,7 +440,7 @@ func (ni *Interface) handleNeighborUpdate(ctx context.Context, nu *netlink.Neigh
 
 	switch nu.Type {
 	case unix.RTM_DELNEIGH:
-		if err := engine.RemoveNeighbor(ctx, ni.fwd, ipToBytes(nu.IP)); err != nil {
+		if err := ni.e.RemoveNeighbor(ctx, ipToBytes(nu.IP)); err != nil {
 			log.Warningf("failed to add neighbor to dataplane: %v", err)
 			return
 		}
@@ -459,7 +456,7 @@ func (ni *Interface) handleNeighborUpdate(ctx context.Context, nu *netlink.Neigh
 			log.Info("skipping neighbor update with no hwaddr")
 			return
 		}
-		if err := engine.AddNeighbor(ctx, ni.fwd, ipToBytes(nu.IP), nu.HardwareAddr); err != nil {
+		if err := ni.e.AddNeighbor(ctx, ipToBytes(nu.IP), nu.HardwareAddr); err != nil {
 			log.Warningf("failed to add neighbor to dataplane: %v", err)
 			return
 		}
@@ -516,17 +513,17 @@ func (ni *Interface) setupPorts(ctx context.Context) error {
 		}
 		ni.idxToName[i.Index] = i.Name
 		ni.idxToName[tap.Index] = tap.Name
-		if err := engine.CreateLocalPort(ctx, ni.fwd, tap.Name, fd); err != nil {
+		if err := ni.e.CreateLocalPort(ctx, tap.Name, fd); err != nil {
 			return fmt.Errorf("failed to create internal port %q: %w", tap.Name, err)
 		}
-		if err := engine.CreateExternalPort(ctx, ni.fwd, i.Name); err != nil {
+		if err := ni.e.CreateExternalPort(ctx, i.Name); err != nil {
 			return fmt.Errorf("failed to create external port %q: %w", i.Name, err)
 		}
 		// Make port only reply to IPs it have
 		if err := os.WriteFile(fmt.Sprintf("/proc/sys/net/ipv4/conf/%s/arp_ignore", i.Name), []byte("2"), 0600); err != nil {
 			return fmt.Errorf("failed to set arp_ignore to 2: %v", err)
 		}
-		if err := engine.UpdatePortSrcMAC(ctx, ni.fwd, i.Name, tap.HardwareAddr); err != nil {
+		if err := ni.e.UpdatePortSrcMAC(ctx, i.Name, tap.HardwareAddr); err != nil {
 			return fmt.Errorf("failed to update MAC address for port %q: %w", i.Name, err)
 		}
 		ni.getOrCreateInterface(i.Name).GetOrCreateEthernet().SetHwMacAddress(tap.HardwareAddr.String())
